matrix phases 1-3: one TCC recipe, N instances — every seat builds every arch
Phase 1 — tcc_instances.mk + prefix tooling:
- tools/tcc_instances.mk: the parameterized one-recipe macro (tcc_instance
  name, target define, output object, prefix). One libtcc.c ONE_SOURCE
  compile per backend, same manifest deps, same include seams.
- tcc_prefix_syms.sh renames EVERY defined global (fixes the gen_negf
  leak seen when two instances coexist); browser wasm objects get the
  python twin wasm_prefix_syms.py (GNU nm/objcopy can't read them).
- Makefile/metal.mk duplicate instance rules collapsed into the recipe.

Phase 1b — arm backend in the manifest + N-instance jit face:
- externals.sh tcc manifest += arm-gen/arm-link/arm-asm.
- jit.c: N-instance shims, TARGET_ARM_EABI enum, arena window for N
  instances.

Phase 2 — firmware in-kernel compile (BIOS + UEFI):
- fw_tcc.mk rewritten: FW_TCC_TARGET selects the backend define (x86_64
  native on x64 boards, ARM native on the armv7 board), FW_TCC_LIB_DIR
  feeds tcc_set_lib_path so TCC never searches /usr/lib; freestanding
  config STATIC/!SEMLOCK/!BACKTRACE/!BCHECK; fwinc shadows every system
  header for the recipe's compile.
- firmware object-output path in jit.c (TCC_OUTPUT_OBJ), arena-backed
  FILE* so libtcc's temp-file API works on bare metal.
- UEFI seam: TCC_CFLAGS undefs _WIN32/_WIN64 — the windows triplet
  defines them, tcc.h goes windows.h/io.h and tccrun pulls
  RtlAddFunctionTable/pe_imagebase; keep the Win64 ABI from the triple,
  drop the macros, TCC compiles the same POSIX shape against fwinc and
  emits the same ELF ET_REL objects (one object format, one loader).
- firmware libc shims (lib.c) for TCC: setjmp/longjmp (jmp_buf fixed at
  10x u64 — the 4-byte-long LLP64 read overflowed TCCState on UEFI and
  freed a corrupted pointer), realpath/mkstemp/unlink/fdopen/freopen/
  getenv/environ/getcwd/mprotect/strerror/localtime/ldexpl/strtof/
  strtold/strtoll/strtoull, fwinc headers (fcntl, sys/mman, sys/time,
  math, stdio, stdlib, string, time, unistd) shaped so the vendored
  TCC compiles freestanding against them.
- QEMU in-kernel compile proves: BIOS + UEFI markers (upy jit c tcc
  object).

Phase 3 — matrix proves on every seat:
- unix: arm-eabi cross-compile loop line (upy cross compile arm loop).
- browser: x64+arm cross knobs (upy browser x64+arm cross emit), wasm
  rename pass via wasm_prefix_syms.py.
- ksweep/build-tests: new cross defines; Makefile greps wired.
- upy_guest_prove.py / upy_browser_prove.py / firmware_upy_cdn.py:
  the prove scripts' new sections (lexer-safe: startswith instead of
  slices, ASCII only — MicroPython's lexer rejects non-ASCII comments).

Gates: prove-all EXIT=0 (host, BIOS, UEFI, RV1106 build, upy, browser);
ksweep 84/84; selfhost ALL STAGES PASSED.

+++ port/fwinc/setjmp.h
@@ -1,7 +1,13 @@
 #ifndef PM_METAL_FW_SETJMP_H
 #define PM_METAL_FW_SETJMP_H
 
-typedef unsigned long jmp_buf[10];
+/* 10 * 8 bytes on EVERY seat: unsigned long is LLP64-4 on the UEFI/Win64
+ * triplet, and the register save below writes 10 quadwords (rsp/rip ride
+ * slots 8/9). A 4-byte long made this 40 bytes on UEFI while the x86_64
+ * setjmp still wrote 80 — error_jmp_buf is embedded in TCCState, so the
+ * overflow smashed 40 bytes of neighboring state fields (the crash free
+ * was error_jmp_buf's shadow: a stack address in the high half). */
+typedef unsigned long long jmp_buf[10];
 
 int setjmp(jmp_buf env);
 void longjmp(jmp_buf env, int val);
